// app-tablas/src/app/pages/machines/gastos/type.ts
export interface GastosTaller {
  centerProvider: string;
  delegation: string;
  date: Date;
  provider: string;
  amount: number;
  currency: string;
  materialReference: string;
  descriptionRefMaterial: string;
  quantity: number;
  switchableCost: boolean;
  description: string;
  deliveryNote: string;
  file: string;
}